separate instant class
